requirements.txt
openai
chromadb
requests
beautifulsoup4
python-dotenv
tiktoken
paramiko
